+++ src/preferences/Preferences.swift
@@ -43,6 +43,8 @@ class Preferences {
             "hideStatusIcons": "false",
             "previewFocusedWindow": "false",
             "captureWindowsInBackground": "true",
+            "taab.dock-click.enabled": "true",
+            "taab.dock-hover-previews.enabled": "true",
             "screenRecordingPermissionSkipped": "false",
             "trackpadHapticFeedbackEnabled": "true",
             "settingsWindowShownOnFirstLaunch": "false",